--- a/src/JustEat.StatsD/StatsDMessageFormatter.cs
+++ b/src/JustEat.StatsD/StatsDMessageFormatter.cs
@@ -5,6 +5,71 @@ using System.Text;
 
 namespace JustEat.StatsD
 {
+    internal static class StringBuilderCache
+    {
+        private const int DefaultCapacity = 128;
+        private const int MaxBuilderSize = DefaultCapacity * 3;
+
+        [ThreadStatic]
+        private static StringBuilder t_cachedInstance;
+
+        public static StringBuilder Acquire(int capacity = DefaultCapacity)
+        {
+            if (capacity <= MaxBuilderSize)
+            {
+                StringBuilder sb = t_cachedInstance;
+                if (capacity < DefaultCapacity)
+                {
+                    capacity = DefaultCapacity;
+                }
+                if (sb != null)
+                {
+                    if (capacity <= sb.Capacity)
+                    {
+                        t_cachedInstance = null;
+                        sb.Clear();
+                        return sb;
+                    }
+                }
+            }
+            return new StringBuilder(capacity);
+        }
+
+        public static void Release(StringBuilder sb)
+        {
+            if (sb.Capacity <= MaxBuilderSize)
+            {
+                t_cachedInstance = sb;
+            }
+        }
+
+        public static string GetStringAndRelease(this StringBuilder sb)
+        {
+            string result = sb.ToString();
+            Release(sb);
+            return result;
+        }
+
+        public static StringBuilder AppendDouble(this StringBuilder sb, double value)
+        {
+#if !NETCOREAPP2_1
+            return sb.Append(value.ToString(CultureInfo.InvariantCulture));
+#else
+            return sb.Append(value);
+#endif
+        }
+
+        public static StringBuilder AppendLong(this StringBuilder sb, long value)
+        {
+#if !NETCOREAPP2_1
+            return sb.Append(value.ToString(CultureInfo.InvariantCulture));
+#else
+            return sb.Append(value);
+#endif
+        }
+
+    }
+
     public class StatsDMessageFormatter
     {
         private const double DefaultSampleRate = 1.0;
@@ -18,13 +83,15 @@ namespace JustEat.StatsD
         public StatsDMessageFormatter()
             : this(string.Empty) {}
 
-            public StatsDMessageFormatter(string prefix)
+        public StatsDMessageFormatter(string prefix)
         {
-            _prefix = prefix;
-
-            if (!string.IsNullOrWhiteSpace(_prefix))
+            if (!string.IsNullOrWhiteSpace(prefix))
             {
-                _prefix = _prefix + "."; // if we have something, then append a . to it to make concatenations easy.
+                _prefix = prefix + "."; // if we have something, then append a . to it to make concatenations easy.
+            }
+            else
+            {
+                _prefix = prefix;
             }
         }
 
@@ -37,7 +104,13 @@ namespace JustEat.StatsD
 
         public string Timing(long milliseconds, double sampleRate, string statBucket)
         {
-            var stat = string.Concat(_prefix, statBucket, ":", milliseconds.ToString("d", InvariantCulture), "|ms");
+            var stat = StringBuilderCache.Acquire()
+                .Append(_prefix)
+                .Append(statBucket)
+                .Append(':')
+                .AppendLong(milliseconds)
+                .Append("|ms");
+            
             return Format(sampleRate, stat);
         }
 
@@ -50,7 +123,6 @@ namespace JustEat.StatsD
         {
             return Decrement(magnitude, DefaultSampleRate, statBucket);
         }
-
 
         public string Decrement(long magnitude, double sampleRate, string statBucket)
         {
@@ -92,7 +164,13 @@ namespace JustEat.StatsD
                 return _prefix + statBucket + ":1|c";
             }
 
-            var stat = string.Concat(_prefix, statBucket, ":", magnitude.ToString(InvariantCulture), "|c");
+            var stat = StringBuilderCache.Acquire()
+                .Append(_prefix)
+                .Append(statBucket)
+                .Append(":")
+                .AppendLong(magnitude)
+                .Append("|c");
+
             return Format(sampleRate, stat);
         }
 
@@ -105,27 +183,54 @@ namespace JustEat.StatsD
         {
             return Format(sampleRate, statBuckets.Select(key => string.Format(InvariantCulture, "{0}{1}:{2}|c", _prefix, key, magnitude)).ToArray());
         }
-
+        
         public string Gauge(double magnitude, string statBucket)
         {
-            var stat = string.Format(InvariantCulture, "{0}{1}:{2}|g", _prefix, statBucket, magnitude);
+            var stat = StringBuilderCache.Acquire()
+                .Append(_prefix)
+                .Append(statBucket)
+                .Append(':')
+                .AppendDouble(magnitude)
+                .Append("|g");
+
             return Format(DefaultSampleRate, stat);
         }
+
         public string Gauge(double magnitude, string statBucket, DateTime timestamp)
         {
-            var stat = string.Format(InvariantCulture, "{0}{1}:{2}|g|@{3}", _prefix, statBucket, magnitude, timestamp.AsUnixTime());
+            var stat = StringBuilderCache.Acquire()
+                .Append(_prefix)
+                .Append(statBucket)
+                .Append(':')
+                .AppendDouble(magnitude)
+                .Append("|g|@")
+                .AppendDouble(timestamp.AsUnixTime());
+
             return Format(DefaultSampleRate, stat);
         }
 
         public string Gauge(long magnitude, string statBucket)
         {
-            var stat = string.Format(InvariantCulture, "{0}{1}:{2}|g", _prefix, statBucket, magnitude);
+            var stat = StringBuilderCache.Acquire()
+                .Append(_prefix)
+                .Append(statBucket)
+                .Append(':')
+                .AppendLong(magnitude)
+                .Append("|g");
+
             return Format(DefaultSampleRate, stat);
         }
 
         public string Gauge(long magnitude, string statBucket, DateTime timestamp)
         {
-            var stat = string.Format(InvariantCulture, "{0}{1}:{2}|g|@{3}", _prefix, statBucket, magnitude, timestamp.AsUnixTime());
+            var stat = StringBuilderCache.Acquire()
+                .Append(_prefix)
+                .Append(statBucket)
+                .Append(':')
+                .AppendLong(magnitude)
+                .Append("|g|@")
+                .AppendDouble(timestamp.AsUnixTime());
+
             return Format(DefaultSampleRate, stat);
         }
 
@@ -134,24 +239,25 @@ namespace JustEat.StatsD
             return Increment(name);
         }
 
-        private string Format(double sampleRate, string stat)
+        private static string Format(double sampleRate, StringBuilder stat)
         {
             if (sampleRate >= DefaultSampleRate)
             {
-                return stat;
+                return stat.GetStringAndRelease();
             }
 
             if (Random.NextDouble() <= sampleRate)
             {
-                return string.Format(InvariantCulture, "{0}|@{1:f}", stat, sampleRate);
+                return stat.AppendFormat(InvariantCulture, "|@{0:f}", sampleRate).GetStringAndRelease();
             }
 
+            StringBuilderCache.Release(stat);
             return string.Empty;
         }
 
         private string Format(double sampleRate, params string[] stats)
         {
-            var formatted = new StringBuilder();
+            var formatted = StringBuilderCache.Acquire(stats.Length * 128);
             if (sampleRate < DefaultSampleRate)
             {
                 foreach (var stat in stats)
@@ -170,7 +276,7 @@ namespace JustEat.StatsD
                 }
             }
 
-            return formatted.ToString();
+            return formatted.GetStringAndRelease();
         }
     }
 }