Use stringbuilder pool to format messages
Use .netcore21 behavior to reduce allocations when formatting long and doubles

**Master**

```
BenchmarkDotNet=v0.11.0, OS=Windows 10.0.17134.167 (1803/April2018Update/Redstone4)
Intel Core i5-7200U CPU 2.50GHz (Max: 2.51GHz) (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.302
  [Host]     : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT
```

|                   Method |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------- |----------:|----------:|----------:|-------:|----------:|
|             IncrementBy1 |  29.03 ns | 0.2915 ns | 0.2727 ns | 0.0508 |      80 B |
|            IncrementBy12 | 146.28 ns | 0.5863 ns | 0.4896 ns | 0.1118 |     176 B |
| IncrementBy2WithSampling | 448.12 ns | 5.6813 ns | 4.7441 ns | 0.1702 |     268 B |
|                Decrement | 151.00 ns | 2.5842 ns | 2.2908 ns | 0.1118 |     176 B |
|                    Event |  29.09 ns | 0.1943 ns | 0.1622 ns | 0.0508 |      80 B |
|                    Gauge | 261.00 ns | 5.2157 ns | 8.1202 ns | 0.0863 |     136 B |
|                   Timing | 149.84 ns | 1.0165 ns | 0.9011 ns | 0.1221 |     192 B |



```
BenchmarkDotNet=v0.11.0, OS=Windows 10.0.17134.167 (1803/April2018Update/Redstone4)
Intel Core i5-7200U CPU 2.50GHz (Max: 2.51GHz) (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.302
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
```


|                   Method |      Mean |     Error |    StdDev |    Median |  Gen 0 | Allocated |
|------------------------- |----------:|----------:|----------:|----------:|-------:|----------:|
|             IncrementBy1 |  33.13 ns | 1.2209 ns | 3.4234 ns |  31.96 ns | 0.0508 |      80 B |
|            IncrementBy12 | 120.98 ns | 2.6486 ns | 3.5358 ns | 119.40 ns | 0.1118 |     176 B |
| IncrementBy2WithSampling | 360.55 ns | 5.2366 ns | 4.3728 ns | 359.89 ns | 0.1473 |     232 B |
|                Decrement | 130.35 ns | 2.8636 ns | 2.5385 ns | 129.71 ns | 0.1118 |     176 B |
|                    Event |  31.18 ns | 0.6916 ns | 0.6469 ns |  31.00 ns | 0.0508 |      80 B |
|                    Gauge | 274.28 ns | 5.3762 ns | 4.4894 ns | 272.78 ns | 0.0658 |     104 B |
|                   Timing | 137.27 ns | 2.8233 ns | 3.6710 ns | 136.20 ns | 0.1218 |     192 B |


**This branch**

```
BenchmarkDotNet=v0.11.0, OS=Windows 10.0.17134.167 (1803/April2018Update/Redstone4)
Intel Core i5-7200U CPU 2.50GHz (Max: 2.51GHz) (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.302
  [Host]     : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.7 (CoreCLR 4.6.26328.01, CoreFX 4.6.26403.03), 64bit RyuJIT
```


|                   Method |      Mean |     Error |    StdDev |    Median |  Gen 0 | Allocated |
|------------------------- |----------:|----------:|----------:|----------:|-------:|----------:|
|             IncrementBy1 |  31.34 ns | 0.9319 ns | 0.8261 ns |  31.14 ns | 0.0508 |      80 B |
|            IncrementBy12 | 145.77 ns | 3.1193 ns | 7.4737 ns | 142.52 ns | 0.0710 |     112 B |
| IncrementBy2WithSampling | 381.08 ns | 1.6934 ns | 1.4141 ns | 381.28 ns | 0.0787 |     124 B |
|                Decrement | 144.70 ns | 0.7741 ns | 0.7241 ns | 144.70 ns | 0.0710 |     112 B |
|                    Event |  29.08 ns | 0.5488 ns | 0.5134 ns |  29.07 ns | 0.0508 |      80 B |
|                    Gauge | 136.78 ns | 1.2735 ns | 1.0634 ns | 136.64 ns | 0.0710 |     112 B |
|                   Timing | 147.13 ns | 1.7325 ns | 1.4467 ns | 146.83 ns | 0.0813 |     128 B |


```

BenchmarkDotNet=v0.11.0, OS=Windows 10.0.17134.167 (1803/April2018Update/Redstone4)
Intel Core i5-7200U CPU 2.50GHz (Max: 2.51GHz) (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.302
  [Host]   : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  ShortRun : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
```

|  Method |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|  ------------------------- |----------:|----------:|----------:|-------:|----------:|
|               IncrementBy1 |  37.85 ns | 23.520 ns | 1.3289 ns | 0.0508 |      80 B |
|              IncrementBy12 | 140.49 ns | 63.380 ns | 3.5811 ns | 0.0508 |      80 B |
|   IncrementBy2WithSampling | 324.19 ns | 78.441 ns | 4.4321 ns | 0.0353 |      56 B |
|                  Decrement | 168.17 ns | 99.249 ns | 5.6077 ns | 0.0508 |      80 B |
|                      Event |  38.73 ns |  8.899 ns | 0.5028 ns | 0.0508 |      80 B |
|                      Gauge | 138.31 ns | 54.684 ns | 3.0898 ns | 0.0508 |      80 B |
|                     Timing | 149.60 ns | 24.501 ns | 1.3844 ns | 0.0558 |      88 B |