@dv00d00 2018-07-31
I've removed StringBuilder pooling, allocations are much higher than in master. branch This is what I got with newing up SB with 64 chars capacity:

|  Method |      Mean |      Error |    StdDev |  Gen 0 | Allocated |
|  ------------------------- |----------:|-----------:|----------:|-------:|----------:|
|               IncrementBy1 |  32.44 ns |  12.198 ns | 0.6892 ns | 0.0044 |      80 B |
|              IncrementBy12 | 117.25 ns |  55.892 ns | 3.1580 ns | 0.0169 |     312 B |
|   IncrementBy2WithSampling | 277.64 ns | 108.044 ns | 6.1047 ns | 0.0157 |     288 B |
|                  Decrement | 124.89 ns |  12.839 ns | 0.7254 ns | 0.0172 |     312 B |
|                      Event |  31.03 ns |   5.740 ns | 0.3243 ns | 0.0044 |      80 B |
|                      Gauge | 114.51 ns |  16.219 ns | 0.9164 ns | 0.0172 |     312 B |
|                     Timing | 121.41 ns |   7.762 ns | 0.4385 ns | 0.0179 |     328 B |