@martincostello 2018-07-31
Internally we can have metrics with lengths of more than 64 characters, so a default of 64 would cause every stat send to resize a StringBuilder under-the-hood before calling `ToString()`